src/requirements.txt
# Core dependencies
openai>=1.0.0
langchain>=0.1.0
langgraph>=0.0.20
python-dotenv>=1.0.0
streamlit>=1.30.0

# Web scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
ddgs>=1.0.0  # DuckDuckGo search (renamed from duckduckgo-search)

# Data processing
pydantic>=2.0.0

# Data storage
PyYAML>=6.0.0
openpyxl>=3.1.0

# Logging
colorlog>=6.8.0
